+++ public/index.html
@@ -22,7 +22,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Hjem</title>
+    <title>NVA</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
