+++ src/pages/publication/references_tab/JournalPublicationReferenceForm.tsx
@@ -4,13 +4,12 @@ import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
 import { FormControl, InputLabel, MenuItem, Select, TextField } from '@material-ui/core';
-import CancelIcon from '@material-ui/icons/Cancel';
-import CheckCircleIcon from '@material-ui/icons/CheckCircle';
 
 import { emptyPublisher, PublicationFormsData } from '../../../types/form.types';
 import { JournalPublicationFieldNames, journalPublicationTypes } from '../../../types/references.types';
 import { PublicationTableNumber } from '../../../utils/constants';
 import JournalPublisherRow from './components/JournalPublisherRow';
+import NviValidation from './components/NviValidation';
 import PeerReview from './components/PeerReview';
 import PublicationChannelSearch from './components/PublicationChannelSearch';
 
@@ -19,41 +18,6 @@ const StyledArticleDetail = styled.div`
   grid-template-columns: auto auto auto auto auto auto;
   grid-column-gap: 1rem;
   align-content: center;
-`;
-
-const StyledNviValidation = styled.div`
-  margin-top: 1rem;
-  display: grid;
-  grid-template-columns: 4rem auto;
-  grid-template-areas:
-    'icon header'
-    'icon information';
-  background-color: ${({ theme }) => theme.palette.background.default};
-  padding: 1rem 0;
-`;
-
-const StyledNviHeader = styled.div`
-  font-size: 1.5rem;
-  font-weight: bold;
-  grid-area: header;
-`;
-
-const StyledNviInformation = styled.div`
-  grid-area: information;
-`;
-
-const StyledCheckCircleIcon = styled(CheckCircleIcon)`
-  grid-area: icon;
-  color: green;
-  margin: 1rem;
-  font-size: 2rem;
-`;
-
-const StyledCancelIcon = styled(CancelIcon)`
-  grid-area: icon;
-  color: red;
-  margin: 1rem;
-  font-size: 2rem;
 `;
 
 const StyledLabel = styled.div`
@@ -106,7 +70,7 @@ const JournalPublicationReferenceForm: React.FC = () => {
               clearSearchField={value === emptyPublisher}
               label={t('publication:references.journal')}
               publicationTable={PublicationTableNumber.PUBLICATION_CHANNELS}
-              setValueFunction={value => setFieldValue(name, value ?? emptyPublisher)}
+              setValueFunction={inputValue => setFieldValue(name, inputValue ?? emptyPublisher)}
             />
             {value.title && (
               <JournalPublisherRow
@@ -142,38 +106,9 @@ const JournalPublicationReferenceForm: React.FC = () => {
         </Field>
       </StyledArticleDetail>
       <StyledPeerReview>
-        <Field name={JournalPublicationFieldNames.PEER_REVIEW}>
-          {({ field }: any) => (
-            <PeerReview field={field} label={t('references.peer_review')} setFieldValue={setFieldValue} />
-          )}
-        </Field>
+        <PeerReview fieldName={JournalPublicationFieldNames.PEER_REVIEW} label={t('references.peer_review')} />
       </StyledPeerReview>
-      <StyledNviValidation>
-        <StyledNviHeader>{t('references.nvi_header')}</StyledNviHeader>
-        {isPeerReviewed ? (
-          isRatedJournal ? (
-            <>
-              <StyledCheckCircleIcon />
-              <StyledNviInformation>{t('references.nvi_success')}</StyledNviInformation>
-            </>
-          ) : (
-            <>
-              <StyledCancelIcon />
-              <StyledNviInformation>
-                <div>{t('references.nvi_fail_rated_line1')}</div>
-                <div>{t('references.nvi_fail_rated_line2')}</div>
-              </StyledNviInformation>
-            </>
-          )
-        ) : (
-          <>
-            <StyledCancelIcon />
-            <StyledNviInformation>
-              <div>{t('references.nvi_fail_no_peer_review')}</div>
-            </StyledNviInformation>
-          </>
-        )}
-      </StyledNviValidation>
+      <NviValidation isPeerReviewed={isPeerReviewed} isRated={!!isRatedJournal} />
     </>
   );
 };
