+++ src/pages/publication/references_tab/components/PeerReview.tsx
@@ -1,8 +1,11 @@
+import { Field, FormikProps, useFormikContext } from 'formik';
 import React, { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
 import { FormControl, FormControlLabel, Radio, RadioGroup } from '@material-ui/core';
+
+import { PublicationFormsData } from '../../../../types/form.types';
 
 const StyledLabel = styled.div`
   color: ${({ theme }) => theme.palette.text.primary};
@@ -10,26 +13,31 @@ const StyledLabel = styled.div`
 `;
 
 interface PeerReviewProps {
-  field: any;
+  fieldName: string;
   label: string;
-  setFieldValue: any;
 }
 
-const PeerReview: FC<PeerReviewProps> = ({ field, label, setFieldValue }) => {
+const PeerReview: FC<PeerReviewProps> = ({ fieldName, label }) => {
   const { t } = useTranslation('publication');
 
+  const { setFieldValue }: FormikProps<PublicationFormsData> = useFormikContext();
+
   return (
-    <>
-      <StyledLabel>{label}</StyledLabel>
-      <FormControl>
-        <RadioGroup
-          value={field.value ? 'true' : 'false'}
-          onChange={event => setFieldValue(field.name, event.target.value === 'true')}>
-          <FormControlLabel control={<Radio value="true" />} label={t('references.is_peer_reviewed')} />
-          <FormControlLabel control={<Radio value="false" />} label={t('references.is_not_peer_reviewed')} />
-        </RadioGroup>
-      </FormControl>
-    </>
+    <Field name={fieldName}>
+      {({ field: { name, value } }: any) => (
+        <>
+          <StyledLabel>{label}</StyledLabel>
+          <FormControl>
+            <RadioGroup
+              value={value ? 'true' : 'false'}
+              onChange={event => setFieldValue(name, event.target.value === 'true')}>
+              <FormControlLabel control={<Radio value="true" />} label={t('references.is_peer_reviewed')} />
+              <FormControlLabel control={<Radio value="false" />} label={t('references.is_not_peer_reviewed')} />
+            </RadioGroup>
+          </FormControl>
+        </>
+      )}
+    </Field>
   );
 };
